requirements.txt
bcrypt==3.2.0
cffi==1.14.3
cryptography==3.2
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
PyYAML==5.3.1
six==1.15.0
